\name{factorize}
\alias{factorize}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Generic factorize function from package conf.design
}
\description{
The generic factorize function factorize from package conf.design is 
repeated here for making sure that it does not get overridden by the 
function from package sfsmisc.
}
\usage{
factorize(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{cf. \code{\link[DoE.base]{factorize}}}
  \item{\dots}{cf. \code{\link[DoE.base]{factorize}}}
}
\details{
All methods, apart from the default method, are from package \pkg{DoE.base}. 
The default method is from package \pkg{conf.design}. 
}
\value{
See the detailed documentation in package \pkg{DoE.base}: \code{\link[DoE.base]{factorize}}
}
\author{
Bill Venables for the original, included here by Ulrike Groemping
}
\seealso{Versions of function \code{factorize} in the following packages:\cr
Package \pkg{conf.design}: \code{\link[conf.design]{factorize}}\cr
Package \pkg{DoE.base}: \code{\link[DoE.base]{factorize}}\cr
Package \pkg{sfsmisc}: \code{\link[sfsmisc]{factorize}}
}
\keyword{ array }
\keyword{ design }
