\name{Class design and accessors}
\alias{[.design}
\alias{design}
\alias{desnum}
\alias{run.order}
\alias{design.info}
\alias{ord}

%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Class design and its accessor functions }
\description{
  Convenience functions to quickly access attributes of data frames of the class 
  design that have been generated by functions FrF2, pb or fac.design
}
\usage{
desnum(design)
run.order(design)
design.info(design)
design
\method{[}{design}(x, i, j, drop.attr = TRUE, drop = FALSE)
ord(matrix, decreasing=FALSE)
}
\arguments{
  \item{design}{ data frame of class design as generated by \code{FrF2}, \code{pb}, \code{fac.design},
       or the above extractor function for class \code{design}. The structures of the respective design 
       objects are described in the value sections of the functions that create them. Generally, all 
       class \code{design} objects are lists with entries \code{desnum}, \code{run.order}, and \code{design.info}.
       The differences lie in the details of \code{design.info}.}
  \item{x}{ data frame of class design as generated by \code{FrF2}, \code{pb}, or \code{fac.design} }
  \item{i}{ indices for subsetting rows}
  \item{j}{ indices for subsetting columns}
  \item{drop.attr}{ logical, controls whether or not attributes are dropped; 
       if TRUE, the result is no longer of class \code{design}, and all special design attributes are dropped;
       otherwise, the design attributes are adjusted to reflect the subsetting result}
  \item{drop}{ logical that controls dropping of dimensions in the Extract function for 
       data.frame objects, which is called by the method for class design}
  \item{matrix}{matrix, data frame or also object of class design that is to be ordered column by column}
  \item{decreasing}{logical, indicates whether decreasing order or not (increasing is default)}
}
\details{
    Items of class \code{design} are data frames with attributes. 
    The functions \code{desnum}, \code{run.order} and \code{design.info} extract 
    the respective attribute, i.e. e.g. function design.info 
    extracts the design information for the design, which depends on the way the 
    design was obtained. Detailed information on the structure of \code{design.info} 
    can be found in the value sections of the respective functions.
    
    The extractor method subsets the design, taking care of the attributes accordingly.
    
    Note that the information attached to objects of class \code{design} is currently only 
    available for usage by the user, but is not (yet) used by the analysis facilities in the 
    package!
}
\value{
    \item{desnum}{returns a numeric matrix,}
    \item{run.order}{returns a 3-column data frame with standard and actual run order 
         as well as a run order with replication identifiers attached,}
    \item{design.info}{returns a list with first element \code{type}, 
        further info on the design, and some options of the design call regarding 
        randomization and replication}
    \item{}{The extractor function returns a class design object with modified attributes 
        or a data frame without special attributes, depending on the situation. 
        
        If \code{j} is given, the function always returns a data frame without 
        special attributes, even if \code{drop.attr=FALSE} or \code{j=1:ncol(design)}.\cr
        If only \code{i} is given, 
        the default option \code{drop.attr=TRUE} also returns a 
        data frame without attributes.\cr
        Exception: Even for \code{drop.attr=TRUE}, if \code{i} is a permutation of 
        the row numbers or a logical vector with all elements TRUE, 
        the attributes are preserved, and attributes \code{run.order} 
        and \code{desnum} are reordered along with the design, if necessary.\cr
        If \code{drop.attr=FALSE} and \code{j} is empty, 
        the function returns an object of class design with rows of attributes \code{run.order} 
        and \code{desnum} selected in line with those of the design itself. 
        In this case, the new \code{design.info} attribute is a list with entries 
        \itemize{
        \item{type}{resolving to \dQuote{subset of design}, }
        \item{subset.rows}{ a numeric or logical vector with the selected rows, and}
        \item{orig.design.info}{ which contains the original design.info attribute.}
        }}
    \item{ord}{returns an index vector that orders the matrix or data frame; 
        for example, \code{design[ord(design),]} orders the design in increasing order with respect to the first,
        then the second etc. factor.}
}
\author{ Ulrike Groemping }
\note{ The package is currently subject to intensive development; most key functionality is now included. 
Some changes to input and output structures may still occur.}
\seealso{ ~~See Also \code{\link{FrF2}}, \code{\link{pb}}, \code{\link{fac.design}} }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ array }
\keyword{ design }% __ONLY ONE__ keyword per line
